pyproject.toml
[tool.poetry]
name = "genshin-helper"
version = "0.0.2"
description = ""
authors = ["FawenYo <sky80838@gmail.com>"]
readme = "README.md"
packages = [{include = "genshin_helper"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
requests = "^2.28.1"
line-bot-sdk = "^2.3.0"
uvicorn = "^0.20.0"
uvloop = "^0.17.0"
aiofiles = "^22.1.0"
Jinja2 = "^3.1.2"
redis = "^4.3.5"
python-dotenv = "^0.21.0"
rich = "^12.6.0"
tqdm = "^4.64.1"
genshin = "^1.5.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
